backend/requirements.txt
fastapi
uvicorn
typing-extensions
python-multipart
Pillow
Logger
torch
transformers
faiss-cpu
tqdm
sqlalchemy
